requirements.txt
# Core dependencies for Galling PINN Model

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Data processing
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Machine learning utilities
scikit-learn>=1.3.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=6.5.4
ipykernel>=6.23.0
